# openapi/agent-gateway-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "quant-trade-agent Gateway",
    "version": "v1"
  },
  "paths": {
...
    "/v1/analyze": {
      "post": {
        "summary": "Analyze",
        "operationId": "analyze_v1_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/chat": {
      "post": {
        "summary": "Run Chat",
        "operationId": "run_chat_v1_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/enrich/headlines/sentiment": {
      "post": {
        "summary": "Enrich Headline Sentiment",
        "operationId": "enrich_headline_sentiment_v1_enrich_headlines_sentiment_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeadlineSentimentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeadlineSentimentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/enrich/sentiment-summary": {
      "post": {
        "summary": "Enrich Sentiment Summary",
        "operationId": "enrich_sentiment_summary_v1_enrich_sentiment_summary_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SentimentSummaryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SentimentSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/narratives/gap": {
      "post": {
        "summary": "Generate Gap Narrative",
        "operationId": "generate_gap_narrative_v1_narratives_gap_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GapNarrativeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GapNarrativeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/translate": {
      "post": {
        "summary": "Translate",
        "operationId": "translate_v1_translate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranslationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/interpret/wish": {
      "post": {
        "summary": "Interpret Wish",
        "operationId": "interpret_wish_v1_interpret_wish_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WishInterpretationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WishInterpretationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/review/code": {
      "post": {
        "summary": "Review Code",
        "operationId": "review_code_v1_review_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/review/respond": {
      "post": {
        "summary": "Respond To Review",
        "operationId": "respond_to_review_v1_review_respond_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRespondRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewRespondResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "const": "9984.T",
            "title": "Symbol",
            "default": "9984.T"
          },
          "question": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "mode": {
            "type": "string",
            "const": "standard",
            "title": "Mode",
            "default": "standard"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AnalyzeRequest"
      },
      "AnalyzeResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "symbol": {
            "type": "string",
            "const": "9984.T",
            "title": "Symbol"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "facts": {
            "additionalProperties": true,
            "type": "object",
            "title": "Facts"
          },
          "analysis": {
            "$ref": "#/components/schemas/AnalysisDetails"
          },
          "decision": {
            "$ref": "#/components/schemas/AnalysisDecision"
          },
          "provenance": {
            "$ref": "#/components/schemas/AnalysisProvenance"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "symbol",
          "as_of",
          "facts",
          "analysis",
          "decision",
          "provenance",
          "warnings"
        ],
        "title": "AnalyzeResponse"
      },
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/HistoryMessage"
            },
            "type": "array",
            "title": "History"
          },
          "symbol": {
            "type": "string",
            "const": "9984.T",
            "title": "Symbol",
            "default": "9984.T"
          },
          "allow_expensive_tools": {
            "type": "boolean",
            "title": "Allow Expensive Tools",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "provider": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Model"
          },
          "tools": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tools"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "answer",
          "provider",
          "model",
          "tools"
        ],
        "title": "ChatResponse"
      },
      "CodeReviewRequest": {
        "properties": {
          "diff": {
            "type": "string",
            "maxLength": 120000,
            "minLength": 1,
            "title": "Diff"
          },
          "project_context": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "diff"
        ],
        "title": "CodeReviewRequest"
      },
      "CodeReviewResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "contract_version": {
            "type": "string",
            "const": "v1",
            "title": "Contract Version",
            "default": "v1"
          },
          "review": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Review"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "LGTM",
              "NEEDS_CHANGES"
            ],
            "title": "Verdict"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "review",
          "verdict",
          "provenance",
          "warnings"
        ],
        "title": "CodeReviewResponse"
      },
...
      "GapNarrativeRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "const": "9984.T",
            "title": "Symbol",
            "default": "9984.T"
          },
          "gap_pct": {
            "type": "number",
            "title": "Gap Pct"
          },
          "headlines": {
            "items": {
              "$ref": "#/components/schemas/GapHeadline"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Headlines"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gap_pct",
          "headlines"
        ],
        "title": "GapNarrativeRequest"
      },
      "GapNarrativeResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "contract_version": {
            "type": "string",
            "const": "v1",
            "title": "Contract Version",
            "default": "v1"
          },
          "narrative": {
            "type": "string",
            "maxLength": 60,
            "minLength": 1,
            "title": "Narrative"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "narrative",
          "provenance",
          "warnings"
        ],
        "title": "GapNarrativeResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HeadlineSentimentRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "const": "9984.T",
            "title": "Symbol",
            "default": "9984.T"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/HeadlineItem"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "HeadlineSentimentRequest"
      },
      "HeadlineSentimentResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "contract_version": {
            "type": "string",
            "const": "v1",
            "title": "Contract Version",
            "default": "v1"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/HeadlineScore"
            },
            "type": "array",
            "title": "Scores"
          },
          "missing_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Missing Ids"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "scores",
          "missing_ids",
          "provenance",
          "warnings"
        ],
        "title": "HeadlineSentimentResponse"
      },
...
      "ReviewRespondRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          },
          "context": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ReviewRespondRequest"
      },
...
      "SentimentSummaryRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "const": "9984.T",
            "title": "Symbol",
            "default": "9984.T"
          },
          "headlines": {
            "items": {
              "$ref": "#/components/schemas/SentimentHeadline"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Headlines"
          },
          "price_context": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Price Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "headlines",
          "price_context"
        ],
        "title": "SentimentSummaryRequest"
      },
      "SentimentSummaryResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "contract_version": {
            "type": "string",
            "const": "v1",
            "title": "Contract Version",
            "default": "v1"
          },
          "analysis": {
            "$ref": "#/components/schemas/SentimentAnalysis"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "analysis",
          "provenance",
          "warnings"
        ],
        "title": "SentimentSummaryResponse"
      },
      "TranslationRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Text"
          },
          "source_language": {
            "type": "string",
            "maxLength": 16,
            "minLength": 2,
            "title": "Source Language"
          },
          "target_language": {
            "type": "string",
            "const": "zh-CN",
            "title": "Target Language",
            "default": "zh-CN"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "source_language"
        ],
        "title": "TranslationRequest"
      },
      "TranslationResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "contract_version": {
            "type": "string",
            "const": "v1",
            "title": "Contract Version",
            "default": "v1"
          },
          "translated": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Translated"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "translated",
          "provenance",
          "warnings"
        ],
        "title": "TranslationResponse"
      },
...
      "WishInterpretationRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 8000,
            "minLength": 1,
            "title": "Message"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/HistoryMessage"
            },
            "type": "array",
            "maxItems": 20,
            "title": "History"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "WishInterpretationRequest"
      },
      "WishInterpretationResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "contract_version": {
            "type": "string",
            "const": "v1",
            "title": "Contract Version",
            "default": "v1"
          },
          "reply": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reply"
          },
          "wish": {
            "$ref": "#/components/schemas/WishDetails"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "reply",
          "wish",
          "provenance",
          "warnings"
        ],
        "title": "WishInterpretationResponse"
      }
    }
  }
}